Refactor knowledge base endpoints into modular router structure
